lib.customisation: add self argument handling to makeOverridable
The `makeOverridable` function will now check for a `_autoSelf` parameter
in the given function arguments and use this parameter to pass the
result of calling the given function as the `self` parameter.

Doing this recursive call in `makeOverridable` ensures that calls to
`overrideAttrs` will also update the `self` reference and allow using
this argument to reliably return self references for example in
`passthru`.