lib.customisation: add overrideWithSelf wrapper
The `overrideWithSelf` wrapper allows to mark a function as
self-referential. The `makeOverridable` function will use this
information and call the given function with the `self` argument set to
the result of this function call (with `lib.fix`).

Doing this recursive call in `makeOverridable` ensures that calls to
`overrideAttrs` will also update the `self` reference and allow using
this argument to reliably return `self` references for example in
`passthru`.