openiscsi: cleanup build and add withSystemd option
Add a withSystemd flag to optionally disable systemd support. This
reduces the runtime closure a lot, which is nice when using the package
in an initrd.

Because of build issues when adding the new flag this also cleans up the
whole build process and removes old workarounds. The build still has to
use absolute path replacements all over the place, because they are
hardcoded in the source.